fixed warnings
Former-commit-id: 4e1a486e724fc5d96028a7dee015a78798aeadbf
Former-commit-id: 7fb5a1258cc2a6b70f635a81d1abde817215ac56

+++ Ignite-Core/include/Ignite/Renderer/Light.h
@@ -3,16 +3,16 @@
 
 namespace Ignite
 {
-	constexpr size_t MAX_LIGHTS{ 10 };
+	constexpr uint32_t MAX_LIGHTS{ 10 };
 	
 	struct Light
 	{
 		alignas(16) glm::vec4 Position{1}; //The W is 0 it is a directional light
 		alignas(16) glm::vec3 Intensity{1}; //Colour
 		alignas(16) glm::vec3 ConeDirection{0};
-		alignas(4)  float Attenuation{0.1};
-		alignas(4)  float AmbientCoefficient{0};
-		alignas(4)  float coneAngle{10};
+		alignas(4)  float Attenuation{0.1f};
+		alignas(4)  float AmbientCoefficient{0.0f};
+		alignas(4)  float coneAngle{10.0f};
 	};
 
 	struct LightBuffer
